+++ tests/eval/test_eval_decode_guard_ordering.py
@@ -66,7 +66,7 @@ def _spec(tmp_path: Path, enc_name: str) -> RoundSpec:
         )
     ]
     return RoundSpec(
-        leaf_batch_size=1, leaf_build_threads=1,
+        leaf_batch_size=1, amp_dtype="bf16", leaf_build_threads=1,
         round_id=f"guard_order_{enc_name}", step=1,
         candidate_snapshot=str(tmp_path / "candidate.pt"),
         best_snapshot=str(tmp_path / "best.pt"), best_step=None,
