+++ tests/eval/test_eval_selfplay_child_parity.py
@@ -171,7 +171,8 @@ def graph_engine():
     net.eval()
     engine = LocalInferenceEngine(net, torch.device("cpu"), encoding_spec=spec,
                                   fused_graph_caps=_CAPS,
-                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8)
+                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8,
+                                  amp_dtype="bf16")
     try:
         yield engine, spec
     finally:
@@ -592,7 +593,7 @@ def test_infer_batch_ls_refuses_a_dense_spec() -> None:
         torch.nn.Identity(), torch.device("cpu"), encoding_spec=lookup("v6"),
         fused_graph_caps=None,
         inference_batching=None,
-        max_in_flight=0,
+        max_in_flight=0, amp_dtype="bf16",
     )
     try:
         with pytest.raises(NotImplementedError):
@@ -619,7 +620,7 @@ def test_build_candidate_player_closed_match_refuses_an_unknown_representation()
         torch.nn.Identity(), torch.device("cpu"), encoding_spec=lookup("v6"),
         fused_graph_caps=None,
         inference_batching=None,
-        max_in_flight=0,
+        max_in_flight=0, amp_dtype="bf16",
     )
     try:
         with pytest.raises(EvalDecodeUnsupportedError):
@@ -650,7 +651,7 @@ def test_infer_ls_is_the_same_refusal_predicate_as_infer_batch_ls(graph_engine) 
         torch.nn.Identity(), torch.device("cpu"), encoding_spec=lookup("v6"),
         fused_graph_caps=None,
         inference_batching=None,
-        max_in_flight=0,
+        max_in_flight=0, amp_dtype="bf16",
     )
     try:
         with pytest.raises(NotImplementedError) as single:
