+++ tests/selfplay/test_inference_local.py
@@ -148,7 +148,7 @@ class _ArchTrappingNet(torch.nn.Module):
 def _engine_with(model: torch.nn.Module) -> LocalInferenceEngine:
     return LocalInferenceEngine(model, _CPU, encoding_spec=_GRID_SPEC,
                                 fused_graph_caps=None,
-                                inference_batching=None, max_in_flight=0)
+                                inference_batching=None, max_in_flight=0, amp_dtype="bf16")
 
 
 def _graph_engine() -> LocalInferenceEngine:
@@ -166,7 +166,7 @@ def _graph_engine() -> LocalInferenceEngine:
     net.eval()
     return LocalInferenceEngine(net, _CPU, encoding_spec=_GRAPH_SPEC,
                                 fused_graph_caps=_CAPS,
-                                inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8)
+                                inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8, amp_dtype="bf16")
 
 
 # ══ I-01 — the four dense-decode invariants, numerically ═════════════════════════
@@ -358,7 +358,7 @@ def test_dense_engine_constructs_no_graph_server() -> None:
     )
     engine = LocalInferenceEngine(net, _CPU, encoding_spec=_GRID_SPEC,
                                   fused_graph_caps=None,
-                                  inference_batching=None, max_in_flight=0)
+                                  inference_batching=None, max_in_flight=0, amp_dtype="bf16")
     try:
         assert engine._is_graph is False
         assert engine._graph_server is None
@@ -405,7 +405,7 @@ def test_engine_reads_no_arch_attributes_off_the_model() -> None:
     net = _ArchTrappingNet(probs, np.array([0.0, 0.0]))
     engine = LocalInferenceEngine(net, _CPU, encoding_spec=_GRID_SPEC,
                                   fused_graph_caps=None,
-                                  inference_batching=None, max_in_flight=0)
+                                  inference_batching=None, max_in_flight=0, amp_dtype="bf16")
     engine.infer_batch([board])
     engine.infer_batch_per_cluster([board])
     assert net.sniffed == [], f"arch attributes were read off the model: {net.sniffed}"
@@ -424,7 +424,7 @@ def test_graph_model_with_dense_spec_fails_loud() -> None:
     try:
         engine = LocalInferenceEngine(net, _CPU, encoding_spec=_GRID_SPEC,
                                   fused_graph_caps=None,
-                                  inference_batching=None, max_in_flight=0)
+                                  inference_batching=None, max_in_flight=0, amp_dtype="bf16")
         with pytest.raises(Exception) as err:
             engine.infer_batch([Board()])
         assert not isinstance(err.value, AssertionError)
@@ -441,7 +441,7 @@ def test_dense_model_with_graph_spec_fails_loud() -> None:
     )
     engine = LocalInferenceEngine(net, _CPU, encoding_spec=_GRAPH_SPEC,
                                   fused_graph_caps=_CAPS,
-                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8)
+                                  inference_batching=InferenceBatchingSpec(inference_batch_size=64, inference_max_wait_ms=10), max_in_flight=8, amp_dtype="bf16")
     try:
         board = Board()
         board.apply_move(0, 0)
